+++ service/db/scraper.go
@@ -25,27 +25,12 @@ func TwitterFanart() {
 	scraper.SetProxy(config.MultiTOR)
 	scraper.SetSearchMode(twitterscraper.SearchLatest)
 	for _, Group := range database.GetGroups() {
-		var wg sync.WaitGroup
-		for _, Member := range database.GetMembers(Group.ID) {
-			wg.Add(1)
-			go func(wg *sync.WaitGroup, Member database.Member, Group database.Group) {
-				defer wg.Done()
-				if Member.TwitterHashtags != "" {
-					if Member.Name != "Kaichou" {
-						Newfanart := twitter.TwitterFanart{
-							Member:  Member,
-							Limit:   100,
-							Group:   Group,
-							Scraper: scraper,
-						}
-						Newfanart.CurlTwitter()
-					}
-				} else {
-					log.Info(Member.EnName + " don't have twitter hashtag")
-				}
-			}(&wg, Member, Group)
+		_, err := twitter.CreatePayload(database.GetHashtag(Group.ID), Group, scraper, 100)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"Group": Group.NameGroup,
+			}).Error(err)
 		}
-		wg.Wait()
 	}
 }