+++ tools/config/config.go
@@ -54,6 +54,7 @@ const (
 	VtubersData       = "https://go-simp.human-z.tech"
 	RandomSleep       = 404
 	ChannelPermission = 16
+	FanartLimit       = 20
 )
 
 type ConfigFile struct {